move @sortablejs/core to sortablejs
