--- packages/core/package.json
@@ -1,32 +1,0 @@
-{
-	"name": "@sortablejs/core",
-	"version": "1.10.2",
-	"description": "JavaScript library for reorderable drag-and-drop lists on modern browsers and touch devices. No jQuery required. Supports Meteor, AngularJS, React, Polymer, Vue, Knockout and any CSS library, e.g. Bootstrap.",
-	"main": "./Sortable.js",
-	"module": "modular/sortable.esm.js",
-	"maintainers": [
-		"Konstantin Lebedev <ibnRubaXa@gmail.com>",
-		"Owen Mills <owen23355@gmail.com>"
-	],
-	"repository": {
-		"type": "git",
-		"url": "git://github.com/SortableJS/Sortable.git"
-	},
-	"files": [
-		"Sortable.js",
-		"Sortable.min.js",
-		"modular/"
-	],
-	"keywords": [
-		"sortable",
-		"reorder",
-		"drag",
-		"meteor",
-		"angular",
-		"ng-sortable",
-		"react",
-		"vue",
-		"mixin"
-	],
-	"license": "MIT"
-}
