+++ packages/sortablejs/src/Animation.js
