+++ packages/sortablejs/src/EventDispatcher.js
