+++ packages/sortablejs/src/PluginManager.js
